made Timer file
